% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcfuns-package.R
\docType{package}
\name{xcfuns-package}
\alias{xcfuns}
\alias{xcfuns-package}
\title{xcfuns: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
